refactor(ui): modularize shared build script
